refactor: split compiler helpers
