+++ compile/go/compiler.go
@@ -3,7 +3,6 @@ package gocode
 import (
 	"bytes"
 	"fmt"
-	"reflect"
 	"sort"
 	"strings"
 
@@ -1851,434 +1850,6 @@ func (c *Compiler) compileFunExpr(fn *parser.FunExpr) (string, error) {
 	return "func(" + strings.Join(params, ", ") + ") " + retType + " {\n" + body + "}", nil
 }
 
-// --- Helpers ---
-
-func (c *Compiler) writeln(s string) {
-	c.writeIndent()
-	c.buf.WriteString(s)
-	c.buf.WriteByte('\n')
-}
-
-func (c *Compiler) writeIndent() {
-	for i := 0; i < c.indent; i++ {
-		c.buf.WriteByte('\t')
-	}
-}
-
-func indentBlock(s string, depth int) string {
-	if s == "" {
-		return s
-	}
-	prefix := strings.Repeat("\t", depth)
-	lines := strings.Split(strings.TrimRight(s, "\n"), "\n")
-	for i, line := range lines {
-		lines[i] = prefix + line
-	}
-	return strings.Join(lines, "\n") + "\n"
-}
-
-func sanitizeName(name string) string {
-	var b strings.Builder
-	for i, r := range name {
-		if r == '_' || ('0' <= r && r <= '9' && i > 0) || ('A' <= r && r <= 'Z') || ('a' <= r && r <= 'z') {
-			b.WriteRune(r)
-		} else if r > 127 {
-			b.WriteRune('_')
-		} else {
-			b.WriteRune('_')
-		}
-	}
-	if b.Len() == 0 || !((b.String()[0] >= 'A' && b.String()[0] <= 'Z') || (b.String()[0] >= 'a' && b.String()[0] <= 'z') || b.String()[0] == '_') {
-		return "_" + b.String()
-	}
-	return b.String()
-}
-
-func exportName(name string) string {
-	if name == "" {
-		return ""
-	}
-	runes := []rune(name)
-	if runes[0] >= 'a' && runes[0] <= 'z' {
-		runes[0] = runes[0] - 'a' + 'A'
-	}
-	return string(runes)
-}
-
-func goType(t types.Type) string {
-	switch tt := t.(type) {
-	case types.IntType:
-		return "int"
-	case types.Int64Type:
-		return "int64"
-	case types.FloatType:
-		return "float64"
-	case types.StringType:
-		return "string"
-	case types.BoolType:
-		return "bool"
-	case types.ListType:
-		return "[]" + goType(tt.Elem)
-	case types.MapType:
-		return fmt.Sprintf("map[%s]%s", goType(tt.Key), goType(tt.Value))
-	case types.StructType:
-		return sanitizeName(tt.Name)
-	case types.FuncType:
-		params := make([]string, len(tt.Params))
-		for i, p := range tt.Params {
-			params[i] = goType(p)
-		}
-		ret := goType(tt.Return)
-		if ret == "" || ret == "void" {
-			return fmt.Sprintf("func(%s)", strings.Join(params, ", "))
-		}
-		return fmt.Sprintf("func(%s) %s", strings.Join(params, ", "), ret)
-	case types.VoidType:
-		return ""
-	case types.AnyType:
-		return "any"
-	default:
-		return "any"
-	}
-}
-
-func equalTypes(a, b types.Type) bool {
-	if _, ok := a.(types.AnyType); ok {
-		return true
-	}
-	if _, ok := b.(types.AnyType); ok {
-		return true
-	}
-	if isInt64(a) && (isInt64(b) || isInt(b)) {
-		return true
-	}
-	if isInt64(b) && (isInt64(a) || isInt(a)) {
-		return true
-	}
-	if isInt(a) && isInt(b) {
-		return true
-	}
-	return reflect.DeepEqual(a, b)
-}
-
-func isInt64(t types.Type) bool {
-	_, ok := t.(types.Int64Type)
-	return ok
-}
-
-func isInt(t types.Type) bool {
-	_, ok := t.(types.IntType)
-	return ok
-}
-
-func isFloat(t types.Type) bool {
-	_, ok := t.(types.FloatType)
-	return ok
-}
-
-func isString(t types.Type) bool {
-	_, ok := t.(types.StringType)
-	return ok
-}
-
-func isAny(t types.Type) bool {
-	_, ok := t.(types.AnyType)
-	return ok
-}
-
-func resolveTypeRef(t *parser.TypeRef) types.Type {
-	if t == nil {
-		return types.AnyType{}
-	}
-	if t.Fun != nil {
-		params := make([]types.Type, len(t.Fun.Params))
-		for i, p := range t.Fun.Params {
-			params[i] = resolveTypeRef(p)
-		}
-		var ret types.Type = types.VoidType{}
-		if t.Fun.Return != nil {
-			ret = resolveTypeRef(t.Fun.Return)
-		}
-		return types.FuncType{Params: params, Return: ret}
-	}
-	if t.Generic != nil {
-		name := t.Generic.Name
-		args := t.Generic.Args
-		switch name {
-		case "list":
-			if len(args) == 1 {
-				return types.ListType{Elem: resolveTypeRef(args[0])}
-			}
-		case "map":
-			if len(args) == 2 {
-				return types.MapType{Key: resolveTypeRef(args[0]), Value: resolveTypeRef(args[1])}
-			}
-		}
-		return types.AnyType{}
-	}
-	if t.Simple != nil {
-		switch *t.Simple {
-		case "int":
-			return types.IntType{}
-		case "float":
-			return types.FloatType{}
-		case "string":
-			return types.StringType{}
-		case "bool":
-			return types.BoolType{}
-		default:
-			return types.AnyType{}
-		}
-	}
-	return types.AnyType{}
-}
-
-func (c *Compiler) inferExprType(e *parser.Expr) types.Type {
-	if e == nil {
-		return types.AnyType{}
-	}
-	return c.inferBinaryType(e.Binary)
-}
-
-func (c *Compiler) inferBinaryType(b *parser.BinaryExpr) types.Type {
-	if b == nil {
-		return types.AnyType{}
-	}
-	t := c.inferUnaryType(b.Left)
-	for _, op := range b.Right {
-		rt := c.inferPostfixType(op.Right)
-		switch op.Op {
-		case "+", "-", "*", "/", "%":
-			if isInt64(t) {
-				if isInt64(rt) || isInt(rt) {
-					t = types.Int64Type{}
-					continue
-				}
-			}
-			if _, ok := t.(types.IntType); ok {
-				if _, ok := rt.(types.IntType); ok {
-					t = types.IntType{}
-					continue
-				}
-			}
-			if _, ok := t.(types.FloatType); ok {
-				if _, ok := rt.(types.FloatType); ok {
-					t = types.FloatType{}
-					continue
-				}
-			}
-			if op.Op == "+" {
-				if _, ok := t.(types.StringType); ok {
-					if _, ok := rt.(types.StringType); ok {
-						t = types.StringType{}
-						continue
-					}
-				}
-			}
-			t = types.AnyType{}
-		case "==", "!=", "<", "<=", ">", ">=":
-			t = types.BoolType{}
-		default:
-			t = types.AnyType{}
-		}
-	}
-	return t
-}
-
-func (c *Compiler) inferUnaryType(u *parser.Unary) types.Type {
-	if u == nil {
-		return types.AnyType{}
-	}
-	return c.inferPostfixType(u.Value)
-}
-
-func (c *Compiler) inferPostfixType(p *parser.PostfixExpr) types.Type {
-	if p == nil {
-		return types.AnyType{}
-	}
-	t := c.inferPrimaryType(p.Target)
-	for _, op := range p.Ops {
-		if op.Index != nil && op.Index.Colon == nil {
-			switch tt := t.(type) {
-			case types.ListType:
-				t = tt.Elem
-			case types.MapType:
-				t = tt.Value
-			case types.StringType:
-				t = types.StringType{}
-			default:
-				t = types.AnyType{}
-			}
-		} else if op.Index != nil {
-			switch tt := t.(type) {
-			case types.ListType:
-				t = tt
-			case types.StringType:
-				t = types.StringType{}
-			default:
-				t = types.AnyType{}
-			}
-		} else if op.Call != nil {
-			if ft, ok := t.(types.FuncType); ok {
-				t = ft.Return
-			} else {
-				t = types.AnyType{}
-			}
-		} else if op.Cast != nil {
-			t = resolveTypeRef(op.Cast.Type)
-		}
-	}
-	return t
-}
-
-func (c *Compiler) inferPrimaryType(p *parser.Primary) types.Type {
-	if p == nil {
-		return types.AnyType{}
-	}
-	switch {
-	case p.Lit != nil:
-		switch {
-		case p.Lit.Int != nil:
-			return types.IntType{}
-		case p.Lit.Float != nil:
-			return types.FloatType{}
-		case p.Lit.Str != nil:
-			return types.StringType{}
-		case p.Lit.Bool != nil:
-			return types.BoolType{}
-		}
-	case p.Selector != nil:
-		if c.env != nil {
-			if t, err := c.env.GetVar(p.Selector.Root); err == nil {
-				if len(p.Selector.Tail) == 0 {
-					return t
-				}
-				st, ok := t.(types.StructType)
-				if ok {
-					cur := st
-					for idx, field := range p.Selector.Tail {
-						ft, ok := cur.Fields[field]
-						if !ok {
-							return types.AnyType{}
-						}
-						if idx == len(p.Selector.Tail)-1 {
-							return ft
-						}
-						if next, ok := ft.(types.StructType); ok {
-							cur = next
-						} else {
-							return types.AnyType{}
-						}
-					}
-				}
-			}
-		}
-		return types.AnyType{}
-	case p.Struct != nil:
-		if c.env != nil {
-			if st, ok := c.env.GetStruct(p.Struct.Name); ok {
-				return st
-			}
-		}
-		return types.AnyType{}
-	case p.Call != nil:
-		switch p.Call.Func {
-		case "len":
-			return types.IntType{}
-		case "now":
-			return types.Int64Type{}
-		default:
-			if c.env != nil {
-				if t, err := c.env.GetVar(p.Call.Func); err == nil {
-					if ft, ok := t.(types.FuncType); ok {
-						return ft.Return
-					}
-				}
-			}
-			return types.AnyType{}
-		}
-	case p.Group != nil:
-		return c.inferExprType(p.Group)
-	case p.List != nil:
-		var elemType types.Type = types.AnyType{}
-		if len(p.List.Elems) > 0 {
-			elemType = c.inferExprType(p.List.Elems[0])
-			for _, e := range p.List.Elems[1:] {
-				t := c.inferExprType(e)
-				if !equalTypes(elemType, t) {
-					elemType = types.AnyType{}
-					break
-				}
-			}
-		}
-		return types.ListType{Elem: elemType}
-	case p.Query != nil:
-		elem := c.inferExprType(p.Query.Select)
-		return types.ListType{Elem: elem}
-	case p.Map != nil:
-		var keyType types.Type = types.AnyType{}
-		var valType types.Type = types.AnyType{}
-		if len(p.Map.Items) > 0 {
-			keyType = c.inferExprType(p.Map.Items[0].Key)
-			valType = c.inferExprType(p.Map.Items[0].Value)
-			for _, it := range p.Map.Items[1:] {
-				kt := c.inferExprType(it.Key)
-				vt := c.inferExprType(it.Value)
-				if !equalTypes(keyType, kt) {
-					keyType = types.AnyType{}
-				}
-				if !equalTypes(valType, vt) {
-					valType = types.AnyType{}
-				}
-			}
-		}
-		return types.MapType{Key: keyType, Value: valType}
-	case p.Match != nil:
-		var rType types.Type
-		for _, cs := range p.Match.Cases {
-			t := c.inferExprType(cs.Result)
-			if rType == nil {
-				rType = t
-			} else if !equalTypes(rType, t) {
-				rType = types.AnyType{}
-			}
-		}
-		if rType == nil {
-			rType = types.AnyType{}
-		}
-		return rType
-	}
-	return types.AnyType{}
-}
-
-func resultType(op string, left, right types.Type) types.Type {
-	switch op {
-	case "+", "-", "*", "/", "%":
-		if _, ok := left.(types.IntType); ok {
-			if _, ok := right.(types.IntType); ok {
-				return types.IntType{}
-			}
-		}
-		if _, ok := left.(types.FloatType); ok {
-			if _, ok := right.(types.FloatType); ok {
-				return types.FloatType{}
-			}
-		}
-		if op == "+" {
-			if _, ok := left.(types.StringType); ok {
-				if _, ok := right.(types.StringType); ok {
-					return types.StringType{}
-				}
-			}
-		}
-		return types.AnyType{}
-	case "==", "!=", "<", "<=", ">", ">=":
-		return types.BoolType{}
-	default:
-		return types.AnyType{}
-	}
-}
-
 func hasExpect(p *parser.Program) bool {
 	for _, s := range p.Statements {
 		if containsExpect(s) {
@@ -2559,386 +2130,4 @@ func (c *Compiler) scanPrimaryImports(p *parser.Primary) {
 	case p.Lit != nil:
 		// none
 	}
-}
-
-// Runtime helper functions injected into generated programs.
-const (
-	helperIndex = "func _index(v any, k any) any {\n" +
-		"    switch s := v.(type) {\n" +
-		"    case []any:\n" +
-		"        i, ok := k.(int)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid list index\")\n" +
-		"        }\n" +
-		"        if i < 0 {\n" +
-		"            i += len(s)\n" +
-		"        }\n" +
-		"        if i < 0 || i >= len(s) {\n" +
-		"            panic(\"index out of range\")\n" +
-		"        }\n" +
-		"        return s[i]\n" +
-		"    case []int:\n" +
-		"        i, ok := k.(int)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid list index\")\n" +
-		"        }\n" +
-		"        if i < 0 {\n" +
-		"            i += len(s)\n" +
-		"        }\n" +
-		"        if i < 0 || i >= len(s) {\n" +
-		"            panic(\"index out of range\")\n" +
-		"        }\n" +
-		"        return s[i]\n" +
-		"    case []float64:\n" +
-		"        i, ok := k.(int)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid list index\")\n" +
-		"        }\n" +
-		"        if i < 0 {\n" +
-		"            i += len(s)\n" +
-		"        }\n" +
-		"        if i < 0 || i >= len(s) {\n" +
-		"            panic(\"index out of range\")\n" +
-		"        }\n" +
-		"        return s[i]\n" +
-		"    case []string:\n" +
-		"        i, ok := k.(int)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid list index\")\n" +
-		"        }\n" +
-		"        if i < 0 {\n" +
-		"            i += len(s)\n" +
-		"        }\n" +
-		"        if i < 0 || i >= len(s) {\n" +
-		"            panic(\"index out of range\")\n" +
-		"        }\n" +
-		"        return s[i]\n" +
-		"    case []bool:\n" +
-		"        i, ok := k.(int)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid list index\")\n" +
-		"        }\n" +
-		"        if i < 0 {\n" +
-		"            i += len(s)\n" +
-		"        }\n" +
-		"        if i < 0 || i >= len(s) {\n" +
-		"            panic(\"index out of range\")\n" +
-		"        }\n" +
-		"        return s[i]\n" +
-		"    case string:\n" +
-		"        i, ok := k.(int)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid string index\")\n" +
-		"        }\n" +
-		"        runes := []rune(s)\n" +
-		"        if i < 0 {\n" +
-		"            i += len(runes)\n" +
-		"        }\n" +
-		"        if i < 0 || i >= len(runes) {\n" +
-		"            panic(\"index out of range\")\n" +
-		"        }\n" +
-		"        return string(runes[i])\n" +
-		"    case map[string]any:\n" +
-		"        ks, ok := k.(string)\n" +
-		"        if !ok {\n" +
-		"            panic(\"invalid map key\")\n" +
-		"        }\n" +
-		"        return s[ks]\n" +
-		"    default:\n" +
-		"        panic(\"invalid index target\")\n" +
-		"    }\n" +
-		"}\n"
-
-	helperIndexString = "func _indexString(s string, i int) string {\n" +
-		"    runes := []rune(s)\n" +
-		"    if i < 0 {\n" +
-		"        i += len(runes)\n" +
-		"    }\n" +
-		"    if i < 0 || i >= len(runes) {\n" +
-		"        panic(\"index out of range\")\n" +
-		"    }\n" +
-		"    return string(runes[i])\n" +
-		"}\n"
-
-	helperIter = "func _iter(v any) []any {\n" +
-		"    switch s := v.(type) {\n" +
-		"    case []any:\n" +
-		"        return s\n" +
-		"    case []int:\n" +
-		"        out := make([]any, len(s))\n" +
-		"        for i, v := range s {\n" +
-		"            out[i] = v\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    case []float64:\n" +
-		"        out := make([]any, len(s))\n" +
-		"        for i, v := range s {\n" +
-		"            out[i] = v\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    case []string:\n" +
-		"        out := make([]any, len(s))\n" +
-		"        for i, v := range s {\n" +
-		"            out[i] = v\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    case []bool:\n" +
-		"        out := make([]any, len(s))\n" +
-		"        for i, v := range s {\n" +
-		"            out[i] = v\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    case map[string]any:\n" +
-		"        out := make([]any, 0, len(s))\n" +
-		"        for k := range s {\n" +
-		"            out = append(out, k)\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    case string:\n" +
-		"        runes := []rune(s)\n" +
-		"        out := make([]any, len(runes))\n" +
-		"        for i, r := range runes {\n" +
-		"            out[i] = string(r)\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    default:\n" +
-		"        return nil\n" +
-		"    }\n" +
-		"}\n"
-
-	helperGenText = "func _genText(prompt string) string {\n" +
-		"    resp, err := llm.Chat(context.Background(), []llm.Message{{Role: \"user\", Content: prompt}})\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    return resp.Message.Content\n" +
-		"}\n"
-
-	helperGenEmbed = "func _genEmbed(text string) []float64 {\n" +
-		"    resp, err := llm.Embed(context.Background(), text)\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    return resp.Vector\n" +
-		"}\n"
-
-	helperGenStruct = "func _genStruct[T any](prompt string) T {\n" +
-		"    resp, err := llm.Chat(context.Background(), []llm.Message{{Role: \"user\", Content: prompt}})\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    var out T\n" +
-		"    if err := json.Unmarshal([]byte(resp.Message.Content), &out); err != nil { panic(err) }\n" +
-		"    return out\n" +
-		"}\n"
-
-	helperFetch = "func _fetch(url string, opts map[string]any) any {\n" +
-		"    method := \"GET\"\n" +
-		"    if opts != nil {\n" +
-		"        if m, ok := opts[\"method\"].(string); ok {\n" +
-		"            method = m\n" +
-		"        }\n" +
-		"    }\n" +
-		"    var body io.Reader\n" +
-		"    if opts != nil {\n" +
-		"        if b, ok := opts[\"body\"]; ok {\n" +
-		"            data, err := json.Marshal(b)\n" +
-		"            if err != nil { panic(err) }\n" +
-		"            body = bytes.NewReader(data)\n" +
-		"        }\n" +
-		"    }\n" +
-		"    u, err := url.Parse(url)\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    if opts != nil {\n" +
-		"        if q, ok := opts[\"query\"]; ok {\n" +
-		"            vals := u.Query()\n" +
-		"            for k, v := range _toAnyMap(q) {\n" +
-		"                vals.Set(k, fmt.Sprint(v))\n" +
-		"            }\n" +
-		"            u.RawQuery = vals.Encode()\n" +
-		"        }\n" +
-		"    }\n" +
-		"    req, err := http.NewRequest(method, u.String(), body)\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    if opts != nil {\n" +
-		"        if hs, ok := opts[\"headers\"]; ok {\n" +
-		"            for k, v := range _toAnyMap(hs) {\n" +
-		"                if s, ok := v.(string); ok {\n" +
-		"                    req.Header.Set(k, s)\n" +
-		"                }\n" +
-		"            }\n" +
-		"        }\n" +
-		"    }\n" +
-		"    client := http.DefaultClient\n" +
-		"    if opts != nil {\n" +
-		"        if t, ok := opts[\"timeout\"]; ok {\n" +
-		"            switch v := t.(type) {\n" +
-		"            case int:\n" +
-		"                client = &http.Client{Timeout: time.Duration(v) * time.Second}\n" +
-		"            case int64:\n" +
-		"                client = &http.Client{Timeout: time.Duration(v) * time.Second}\n" +
-		"            case float64:\n" +
-		"                client = &http.Client{Timeout: time.Duration(v * float64(time.Second))}\n" +
-		"            case float32:\n" +
-		"                client = &http.Client{Timeout: time.Duration(float64(v) * float64(time.Second))}\n" +
-		"            }\n" +
-		"        }\n" +
-		"    }\n" +
-		"    resp, err := client.Do(req)\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    defer resp.Body.Close()\n" +
-		"    data, err := io.ReadAll(resp.Body)\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    var out any\n" +
-		"    if err := json.Unmarshal(data, &out); err != nil { panic(err) }\n" +
-		"    return out\n" +
-		"}\n"
-
-	helperToAnyMap = "func _toAnyMap(m any) map[string]any {\n" +
-		"    switch v := m.(type) {\n" +
-		"    case map[string]any:\n" +
-		"        return v\n" +
-		"    case map[string]string:\n" +
-		"        out := make(map[string]any, len(v))\n" +
-		"        for k, vv := range v {\n" +
-		"            out[k] = vv\n" +
-		"        }\n" +
-		"        return out\n" +
-		"    default:\n" +
-		"        return nil\n" +
-		"    }\n" +
-		"}\n"
-
-	helperCast = "func _cast[T any](v any) T {\n" +
-		"    data, err := json.Marshal(v)\n" +
-		"    if err != nil { panic(err) }\n" +
-		"    var out T\n" +
-		"    if err := json.Unmarshal(data, &out); err != nil { panic(err) }\n" +
-		"    return out\n" +
-		"}\n"
-
-	helperQuery = "type _joinSpec struct { items []any; on func(...any) bool; left bool; right bool }\n" +
-		"type _queryOpts struct { selectFn func(...any) any; where func(...any) bool; sortKey func(...any) any; skip int; take int }\n" +
-		"func _query(src []any, joins []_joinSpec, opts _queryOpts) []any {\n" +
-		"    items := make([][]any, len(src))\n" +
-		"    for i, v := range src { items[i] = []any{v} }\n" +
-		"    for _, j := range joins {\n" +
-		"        joined := [][]any{}\n" +
-		"        if j.right && j.left {\n" +
-		"            matched := make([]bool, len(j.items))\n" +
-		"            for _, left := range items {\n" +
-		"                m := false\n" +
-		"                for ri, right := range j.items {\n" +
-		"                    keep := true\n" +
-		"                    if j.on != nil { args := append(append([]any(nil), left...), right); keep = j.on(args...) }\n" +
-		"                    if !keep { continue }\n" +
-		"                    m = true; matched[ri] = true\n" +
-		"                    joined = append(joined, append(append([]any(nil), left...), right))\n" +
-		"                }\n" +
-		"                if !m { joined = append(joined, append(append([]any(nil), left...), nil)) }\n" +
-		"            }\n" +
-		"            for ri, right := range j.items {\n" +
-		"                if !matched[ri] { undef := make([]any, len(items[0])); joined = append(joined, append(undef, right)) }\n" +
-		"            }\n" +
-		"        } else if j.right {\n" +
-		"            for _, right := range j.items {\n" +
-		"                m := false\n" +
-		"                for _, left := range items {\n" +
-		"                    keep := true\n" +
-		"                    if j.on != nil { args := append(append([]any(nil), left...), right); keep = j.on(args...) }\n" +
-		"                    if !keep { continue }\n" +
-		"                    m = true; joined = append(joined, append(append([]any(nil), left...), right))\n" +
-		"                }\n" +
-		"                if !m { undef := make([]any, len(items[0])); joined = append(joined, append(undef, right)) }\n" +
-		"            }\n" +
-		"        } else {\n" +
-		"            for _, left := range items {\n" +
-		"                m := false\n" +
-		"                for _, right := range j.items {\n" +
-		"                    keep := true\n" +
-		"                    if j.on != nil { args := append(append([]any(nil), left...), right); keep = j.on(args...) }\n" +
-		"                    if !keep { continue }\n" +
-		"                    m = true; joined = append(joined, append(append([]any(nil), left...), right))\n" +
-		"                }\n" +
-		"                if j.left && !m { joined = append(joined, append(append([]any(nil), left...), nil)) }\n" +
-		"            }\n" +
-		"        }\n" +
-		"        items = joined\n" +
-		"    }\n" +
-		"    if opts.where != nil {\n" +
-		"        filtered := [][]any{}\n" +
-		"        for _, r := range items { if opts.where(r...) { filtered = append(filtered, r) } }\n" +
-		"        items = filtered\n" +
-		"    }\n" +
-		"    if opts.sortKey != nil {\n" +
-		"        type pair struct { item []any; key any }\n" +
-		"        pairs := make([]pair, len(items))\n" +
-		"        for i, it := range items { pairs[i] = pair{it, opts.sortKey(it...)} }\n" +
-		"        sort.Slice(pairs, func(i, j int) bool {\n" +
-		"            a, b := pairs[i].key, pairs[j].key\n" +
-		"            switch av := a.(type) {\n" +
-		"            case int:\n" +
-		"                switch bv := b.(type) { case int: return av < bv; case float64: return float64(av) < bv }\n" +
-		"            case float64:\n" +
-		"                switch bv := b.(type) { case int: return av < float64(bv); case float64: return av < bv }\n" +
-		"            case string:\n" +
-		"                bs, _ := b.(string); return av < bs\n" +
-		"            }\n" +
-		"            return fmt.Sprint(a) < fmt.Sprint(b)\n" +
-		"        })\n" +
-		"        for i, p := range pairs { items[i] = p.item }\n" +
-		"    }\n" +
-		"    if opts.skip >= 0 { if opts.skip < len(items) { items = items[opts.skip:] } else { items = [][]any{} } }\n" +
-		"    if opts.take >= 0 { if opts.take < len(items) { items = items[:opts.take] } }\n" +
-		"    res := make([]any, len(items))\n" +
-		"    for i, r := range items { res[i] = opts.selectFn(r...) }\n" +
-		"    return res\n" +
-		"}\n"
-)
-
-var helperMap = map[string]string{
-	"_index":       helperIndex,
-	"_indexString": helperIndexString,
-	"_iter":        helperIter,
-	"_genText":     helperGenText,
-	"_genEmbed":    helperGenEmbed,
-	"_genStruct":   helperGenStruct,
-	"_fetch":       helperFetch,
-	"_toAnyMap":    helperToAnyMap,
-	"_cast":        helperCast,
-	"_query":       helperQuery,
-}
-
-func (c *Compiler) use(name string) {
-	c.helpers[name] = true
-}
-
-func (c *Compiler) emitRuntime() {
-	if len(c.helpers) == 0 {
-		return
-	}
-	names := make([]string, 0, len(c.helpers))
-	for n := range c.helpers {
-		names = append(names, n)
-	}
-	sort.Strings(names)
-	for _, n := range names {
-		c.buf.WriteString(helperMap[n])
-		c.buf.WriteByte('\n')
-	}
-}
-
-func isUnderscoreExpr(e *parser.Expr) bool {
-	if e == nil {
-		return false
-	}
-	if len(e.Binary.Right) != 0 {
-		return false
-	}
-	u := e.Binary.Left
-	if len(u.Ops) != 0 {
-		return false
-	}
-	p := u.Value
-	if len(p.Ops) != 0 {
-		return false
-	}
-	if p.Target.Selector != nil && p.Target.Selector.Root == "_" && len(p.Target.Selector.Tail) == 0 {
-		return true
-	}
-	return false
 }
